Handle `Enumerable` values on `Repository#delete`
A test for `RDF::Enumerable` inputs was added by
https://github.com/ruby-rdf/rdf-spec/pull/39. This adds conformance, and
allows `SPARQL::Client::Repository` to use the new
`Mutable#delete_insert` interface. It does not yet implement an
effecient SPARQL `#delete_insert`.

It may be possible to refactor `#delete_statements` in response to these
changes, to remove the code smells called out in the comment in that
method.