Pass 2.0 transaction specs
`#delete_data` and `#insert_data` can error when given an empty enum for
a statement set. This ran afoul of some specs for transactions.

It's probably best that these methods handle this case, anyway, but this
solution is a bit hacky. There's likely a better solution via the
update_client.

This is submitted as proof of concept, rather than as proposed code.