raise error when no suitable rdf reader is found
Currently sparql-client will silently fail with a nil if no suitable rdf reader is found. This can lead to the awkward situation where a nil is returned for the query and the user has to find out what is causing it. This commit tries to quicken this debugging proces by at least indicating the point of failure by raising a descriptive error.