Fixed problems with the alternative endpoint
The alternative endpoint is set optionally in the update method if an
endpoint is provided in the options. If the next call doesn't provide an
alternative endpoint, it should use the instance's configured endpoint.

Also: a call to the query method may use the make_post_request method.
We should make sure to not keep any previous alternative endpoint.

Note: this commit not only fix the current issues with the alternative
endpoints, it also implements a new functionality: the possibility to
override the configured endpoint when calling the query method.

(cherry picked from commit 82530b90895a1c17fea4d85f08c23cc08e3324d7)